split __init__.py into multiple modules
* Most of the parsing of arguments (and therefore most of the old
__init__.py goes into cli.py. This also conforms to the test_ module
by the same name.
* Exceptions are split out into their own module, a common practice
* A module, compat.py, is added for any PY2 and PY3 fixes
* VCS (git and mercurial) are now in their own module
* __init__.py only contains some meta symbols describing the package
* __main__.py now simply executes the cli.main() and nothing else
* Loggers are now named after the module they reside in, allowing
for simpler debugging. This does alter the tests quite a bit as we
(unfortunately) rely on logging output quite a bit. In particular,
we check for calls to the logging handler (which may or may not end
up being actual logs depending on the log level set). With multiple
handlers (one per module) we cannot mock as easily so we switch to
testfixtures capturing logs.