Treat file-like unnecessary part as non-existing file
It's 'file-like' if it includes a dot, slash or backslash.
Unless of course there is matching part configuration (which
includes the dot/slash/backslash).

Without this change, test 'non_existing_file' fails because
bump2version considers the supplied non-existing path to be an
unnecessary [part] argument and only produces a
DeprecationWarning for it.

While of course we want a proper FileNotFoundError (IOError) if we
supply an argument that looks like a file and can not be found on
disk.

Parametrize the 'non existing file' test to check all cases of dot,
slash, backslash.